refactor isEnabled in separate package configcheck
