+++ comp/otelcol/collector/impl-pipeline/status.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/comp/core/status"
 	"github.com/DataDog/datadog-agent/comp/otelcol/otlp"
+	"github.com/DataDog/datadog-agent/comp/otelcol/otlp/configcheck"
 )
 
 //go:embed status_templates
@@ -28,7 +29,7 @@ func (c *collectorImpl) getStatusInfo() map[string]interface{} {
 
 func (c *collectorImpl) populateStatus(stats map[string]interface{}) {
 	otlpStatus := make(map[string]interface{})
-	otlpIsEnabled := otlp.IsEnabled(c.config)
+	otlpIsEnabled := configcheck.IsEnabled(c.config)
 
 	var otlpCollectorStatus otlp.CollectorStatus
