+++ comp/otelcol/otlp/config.go
@@ -12,12 +12,11 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/mohae/deepcopy"
-	"go.opentelemetry.io/collector/confmap"
 	"go.uber.org/multierr"
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	"github.com/DataDog/datadog-agent/comp/otelcol/otlp/components/exporter/serializerexporter"
+	"github.com/DataDog/datadog-agent/comp/otelcol/otlp/configcheck"
 	coreconfig "github.com/DataDog/datadog-agent/pkg/config/setup"
 	"github.com/DataDog/datadog-agent/pkg/util"
 	"github.com/go-viper/mapstructure/v2"
@@ -31,50 +30,10 @@ func portToUint(v int) (port uint, err error) {
 	return
 }
 
-// readConfigSection from a config.Component object.
-func readConfigSection(cfg config.Reader, section string) *confmap.Conf {
-	// Viper doesn't work well when getting subsections, since it
-	// ignores environment variables and nil-but-present sections.
-	// To work around this, we do the following two steps:
-
-	// Step one works around https://github.com/spf13/viper/issues/819
-	// If we only had the stuff below, the nil sections would be ignored.
-	// We want to take into account nil-but-present sections.
-	//
-	// Furthermore, Viper returns an `interface{}` nil in the case where
-	// `section` is present but empty: e.g. we want to read
-	//	"otlp_config.receiver", but we have
-	//
-	//         otlp_config:
-	//           receiver:
-	//
-	// `GetStringMap` it will fail to cast `interface{}` nil to
-	// `map[string]interface{}` nil; we use `Get` and cast manually.
-	rawVal := cfg.Get(section)
-	stringMap := map[string]interface{}{}
-	if val, ok := rawVal.(map[string]interface{}); ok {
-		// deep copy since `cfg.Get` returns a reference
-		stringMap = deepcopy.Copy(val).(map[string]interface{})
-	}
-
-	// Step two works around https://github.com/spf13/viper/issues/1012
-	// we check every key manually, and if it belongs to the OTLP receiver section,
-	// we set it. We need to do this to account for environment variable values.
-	prefix := section + "."
-	for _, key := range cfg.AllKeysLowercased() {
-		if strings.HasPrefix(key, prefix) && cfg.IsSet(key) {
-			mapKey := strings.ReplaceAll(key[len(prefix):], ".", confmap.KeyDelimiter)
-			// deep copy since `cfg.Get` returns a reference
-			stringMap[mapKey] = deepcopy.Copy(cfg.Get(key))
-		}
-	}
-	return confmap.NewFromStringMap(stringMap)
-}
-
 // FromAgentConfig builds a pipeline configuration from an Agent configuration.
 func FromAgentConfig(cfg config.Reader) (PipelineConfig, error) {
 	var errs []error
-	otlpConfig := readConfigSection(cfg, coreconfig.OTLPReceiverSection)
+	otlpConfig := configcheck.ReadConfigSection(cfg, coreconfig.OTLPReceiverSection)
 	tracePort, err := portToUint(cfg.GetInt(coreconfig.OTLPTracePort))
 	if err != nil {
 		errs = append(errs, fmt.Errorf("internal trace port is invalid: %w", err))
@@ -85,7 +44,7 @@ func FromAgentConfig(cfg config.Reader) (PipelineConfig, error) {
 	if !metricsEnabled && !tracesEnabled && !logsEnabled {
 		errs = append(errs, fmt.Errorf("at least one OTLP signal needs to be enabled"))
 	}
-	metricsConfig := readConfigSection(cfg, coreconfig.OTLPMetrics)
+	metricsConfig := configcheck.ReadConfigSection(cfg, coreconfig.OTLPMetrics)
 	metricsConfigMap := metricsConfig.ToStringMap()
 
 	if _, ok := metricsConfigMap["apm_stats_receiver_addr"]; !ok {
@@ -101,7 +60,7 @@ func FromAgentConfig(cfg config.Reader) (PipelineConfig, error) {
 		errs = append(errs, fmt.Errorf("failed to normalize metrics config: %w", err))
 	}
 
-	debugConfig := readConfigSection(cfg, coreconfig.OTLPDebug)
+	debugConfig := configcheck.ReadConfigSection(cfg, coreconfig.OTLPDebug)
 
 	return PipelineConfig{
 		OTLPReceiverConfig: otlpConfig.ToStringMap(),
@@ -140,29 +99,4 @@ func normalizeMetricsConfig(metricsConfigMap map[string]interface{}, strict bool
 		return nil, err
 	}
 	return mc, nil
-}
-
-// IsEnabled checks if OTLP pipeline is enabled in a given config.
-func IsEnabled(cfg config.Reader) bool {
-	return hasSection(cfg, coreconfig.OTLPReceiverSubSectionKey)
-}
-
-// HasLogsSectionEnabled checks if OTLP logs are explicitly enabled in a given config.
-func HasLogsSectionEnabled(cfg config.Reader) bool {
-	return hasSection(cfg, coreconfig.OTLPLogsEnabled) && cfg.GetBool(coreconfig.OTLPLogsEnabled)
-}
-
-func hasSection(cfg config.Reader, section string) bool {
-	// HACK: We want to mark as enabled if the section is present, even if empty, so that we get errors
-	// from unmarshaling/validation done by the Collector code.
-	//
-	// IsSet won't work here: it will return false if the section is present but empty.
-	// To work around this, we check if the receiver key is present in the string map, which does the 'correct' thing.
-	_, ok := readConfigSection(cfg, coreconfig.OTLPSection).ToStringMap()[section]
-	return ok
-}
-
-// IsDisplayed checks if the OTLP section should be rendered in the Agent
-func IsDisplayed() bool {
-	return true
 }
