+++ pkg/serverless/otlp/otlp.go
@@ -16,6 +16,7 @@ import (
 	"go.opentelemetry.io/collector/otelcol"
 
 	coreOtlp "github.com/DataDog/datadog-agent/comp/otelcol/otlp"
+	"github.com/DataDog/datadog-agent/comp/otelcol/otlp/configcheck"
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
 	"github.com/DataDog/datadog-agent/pkg/serializer"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
@@ -61,7 +62,7 @@ func (o *ServerlessOTLPAgent) Stop() {
 
 // IsEnabled returns true if the OTLP endpoint should be enabled.
 func IsEnabled() bool {
-	return coreOtlp.IsEnabled(pkgconfigsetup.Datadog())
+	return configcheck.IsEnabled(pkgconfigsetup.Datadog())
 }
 
 var (
